reservas-backend/api/permissions.py
from rest_framework.permissions import BasePermission

class IsAdmin(BasePermission):
    def has_permission(self, request, view):
        return bool(request.user and request.user.is_authenticated and hasattr(request.user, "perfil") and request.user.perfil.rol == "admin")

class IsColaborador(BasePermission):
    def has_permission(self, request, view):
        return bool(request.user and request.user.is_authenticated and hasattr(request.user, "perfil") and request.user.perfil.rol == "colaborador")
